Add Ui, TaskList, Storage, Parser classes
Code logic was all placed in Tracker.java
This impedes new developers from understanding
what the Tracker's function is.

It is better to have a segregation of concerns
so that each feature is placed in its own classes

As such 4 new classes are made to accomodate
such functions
