+++ src/main/java/Task.java
@@ -11,12 +11,12 @@ public class Task {
 
     public void mark(boolean b, int taskIndex) {
         String MARK = b ? "marked" : "unmarked"; //use enums if more needed
-        if (Tracker.tasks.isEmpty()) {
+        if (TaskList.tasks.isEmpty()) {
             System.out.println("There's nothing to mark, Yay!");
             return;
         }
         try {
-            if (taskIndex > Tracker.tasks.size() || taskIndex < 0) {
+            if (taskIndex > TaskList.tasks.size() || taskIndex < 0) {
                 System.out.println("You don't have that task silly!");
                 return;
             }
